+++ lib/DAO/settlement_dao.dart
@@ -2,29 +2,29 @@ import 'dart:developer';
 import 'package:intl/intl.dart';
 import 'package:sqflite/sqflite.dart';
 
-class DBSettlement {
+class SettlementDAO {
 
-  DBSettlement();
+  SettlementDAO();
 
-  Future<int> createStm(Database db, String stmId, String stmName) async {
+  Future<int> save(Database db, String stmId, String stmName) async {
     return await db!.rawInsert('INSERT INTO Settlement(settlementId, settlementName, recordDate) VALUES(?, ?, ?)'
         , [stmId, stmName, DateFormat('yyyy-MM-dd HH:mm:ss').format(DateTime.now())]);
   }
 
-  Future<List<Map>> readStm(Database db, String stmId) async {
+  Future<List<Map>> findById(Database db, String stmId) async {
     return await db!.rawQuery('SELECT * FROM Settlement WHERE settlementId = ?', [stmId]);
   }
 
-  Future<int> updateStm(Database db, String newstmName, String stmId) async {
+  Future<int> update(Database db, String newstmName, String stmId) async {
     return await db!.rawUpdate('UPDATE Settlement SET settlementName = ?, recordDate = ? WHERE settlementId = ?'
         , [newstmName, DateFormat('yyyy-MM-dd HH:mm:ss').format(DateTime.now()), stmId]);
   }
 
-  Future<int> deleteStm(Database db, String stmId) async {
+  Future<int> delete(Database db, String stmId) async {
     return await db!.rawDelete('DELETE FROM Settlement WHERE settlementId = ?', [stmId]);
   }
 
-  Future<int> deleteStmTxn(Transaction txn, String stmId) async {
+  Future<int> deleteTxn(Transaction txn, String stmId) async {
     return await txn.rawDelete('DELETE FROM Settlement WHERE settlementId = ?', [stmId]);
   }
